Extract common codec code
