+++ lib/fluent/plugin/out_filter_codec.rb
@@ -1,9 +1,11 @@
 # coding: utf-8
 require "base64"
+require "fluent/plugin/filter_codec_support"
 
 module Fluent
   class FilterCodecOutput < Output
     include Fluent::HandleTagNameMixin
+    include Fluent::FilterCodecSupport
 
     Fluent::Plugin.register_output('filter_codec', self)
 
@@ -57,35 +59,6 @@ module Fluent
       end
 
       record[@output_field] = process_filter(value_in, @codec) || value_in
-    end
-
-    private
-    def process_filter(value, codec)
-      return @codec_functions[codec].call(value)
-    rescue
-      return @error_value
-    end
-
-    def base64_decode(value)
-      return Base64.strict_decode64(value)
-    end
-
-    def base64_encode(value)
-      return Base64.strict_encode64(value)
-    end
-
-    def add_one_padding(str)
-      return (0 != str.length % 4) ? str + '=' : str
-    end
-
-    def urlsafe64_decode(value)
-      # Add up to 2 '=' padding
-      return Base64.urlsafe_decode64(add_one_padding(add_one_padding(value)))
-    end
-
-    def urlsafe64_encode(value)
-      # Remove up to 2 '=' padding
-      return Base64.urlsafe_encode64(value).chomp('=').chomp('=')
     end
   end
 end
